# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/user-roles": {
            "get": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Get all user roles",
                "operationId": "5def8f7f8452aaf9a2b02cdb217d4e1b",
                "responses": {
                    "200": {
                        "description": "List of all user roles"
                    }
                }
            },
            "post": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Create a new user role",
                "operationId": "b19eaa027d6763b45aa3b103cd2c4c0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role_name",
                                    "privilege"
                                ],
                                "properties": {
                                    "role_name": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "privilege": {
                                        "type": "string",
                                        "example": "Full Access"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User role created"
                    }
                }
            }
        },
        "/api/user-roles/{id}": {
            "get": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Get a specific user role",
                "operationId": "df2df80724654054dbddda8082169fb6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user role",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User role details"
                    }
                }
            },
            "put": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Update an existing user role",
                "operationId": "dd1e8c60ea5304801982ecb6549e0da9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user role",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role_name": {
                                        "type": "string",
                                        "example": "user"
                                    },
                                    "privilege": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User role updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Delete a user role",
                "operationId": "41802a14c26b60c2e9afeb09dc7d57f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user role",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User role deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rola bola vymazaná."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}